requirements.txt
MouseInfo==0.1.3
msedge-selenium-tools==3.141.3
Pillow==8.3.2
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
pytweening==1.0.4
selenium==3.141.0
urllib3==1.26.7
